added a toy script to scrape text
this toy script is not perfect but it can scrape text from a pdf page by page. to complete the issue, it is recommended that the toy script gets converted into a proper function and there is some sort of help with formatting the parsed text ... mainly adding spaces where they should be.